% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTDS_crop.R
\name{LTDS_crop}
\alias{LTDS_crop}
\title{LTDS crop}
\usage{
LTDS_crop(ltds_data, trunc)
}
\arguments{
\item{ltds_data}{Raw field data from a full LTDS survey}

\item{trunc}{The truncation distance (half distance between transects)}

\item{save}{Do you want to save the data to the working directory?}
}
\value{
A csv file of truncated ltds data
}
\description{
Crops field data to remove observations outside of truncation distance and outliers
}
\examples{
#Read in csv file showing transect lines
ltds_data <- read_csv("./LTDS_example_data.csv")

#Input raw data into function
new_ltds_data <- ltds_crop(ltds_data = ltds_data, trunc = 5, save = T)

}
