% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTDS_summary.R
\name{LTDS_summary}
\alias{LTDS_summary}
\title{LTDS summary}
\usage{
LTDS_summary(
  effort = "NA",
  nburr = "NA",
  burr.w.mn = "NA",
  burr.w.min = "NA",
  burr.w.max = "NA",
  comm.comm = "NA",
  occ = "NA",
  unocc = "NA",
  unk = "NA",
  site.name = "NA",
  tort.dens = "NA",
  tort.pop = "NA"
)
}
\arguments{
\item{effort}{Total effort in meters from full LTDS survey}

\item{nburr}{Number of burrows detected}

\item{burr.w.mn}{Mean burrow width}

\item{burr.w.min}{Minimum burrow width}

\item{burr.w.max}{Maximum burrow width}

\item{comm.comm}{Most common commensal species}

\item{occ}{Number of occupied burrows discovered}

\item{unocc}{Number of unoccupied burrows discovered}

\item{unk}{Number of unknown burrows discovered}

\item{site.name}{Name of recipient site}

\item{tort.dens}{Tortoise density per acre at recipient site}

\item{tort.pop}{Tortoise population estimate at recipient site}
}
\value{
A PDF report of LTDS data
}
\description{
Uses LTDS field data and estimation results to create a report summary.
}
\examples{
#Read in csv file showing full LTDS survey data
ltds_data <- read_csv("./LTDS_example_data.csv")

#Input raw data into function
new_ltds_data <- ltds_summary(ltds_data = ltds_data, mort_data = mort_data, save = T)

}
