% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survey_effort.R
\name{samp_effort}
\alias{samp_effort}
\title{Survey effort}
\usage{
samp_effort(points, zone, plot)
}
\arguments{
\item{points}{A csv file showing start and end locations for LTDS lines}

\item{zone}{The UTM zone of the study site}

\item{plot}{Do you want the actual lines walked to be plotted?}
}
\value{
Distance in meters showing effort of pilot or full survey
}
\description{
Calculates the effort as distance in meters
}
\examples{
#Read in csv file showing transect lines
start.end <- read_csv("./Start_end_locations.csv")

#Input coordinate file
effort <- samp_effort(start.end, zone = 17, plot = T)
#View resulting effort
effort

}
