% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Full_survey.R
\name{samp_full}
\alias{samp_full}
\title{Full LTDS survey effort}
\usage{
samp_full(effort, site.poly, direction, nburr, cv, zone, plot, save)
}
\arguments{
\item{effort}{Distance in meters walked during the pilot survey}

\item{site.poly}{A shape file used to clip the transect lines to fit the site}

\item{direction}{The cardinal direction of the transects}

\item{nburr}{Number of occupied burrows detected during the pilot survey}

\item{cv}{The target coefficient of variation for full LTDS survey (e.g. 0.17 = 17%)}

\item{zone}{The UTM zone of the shapefile/study site}

\item{plot}{Do you want the transect placement to be plotted against the site?}

\item{save}{Do you want to save the transect lines to the working directory?}
}
\value{
A SpatialLines object delineating additional transect placement
}
\description{
Delineates additional transects needed for full LTDS survey
}
\examples{
#Read in csv file showing transect lines
start.end <- read_csv("./Start_end_locations.csv")

#Input coordinate file
effort <- samp_effort(start.end, zone = 17, plot = T)

#Input resulting pilot effort into function
full_survey <- samp_full(effort, site.poly, direction = "N-S", nburr = 100, cv = 0.17, zone = 17, plot = T, save = T)

#To view the effort needed, run:
gLength(full_survey)


}
