% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site_merge.R
\name{site_merge}
\alias{site_merge}
\title{Site merge}
\usage{
site_merge(shape1, shape2, zone, plot, save)
}
\arguments{
\item{shape1}{The first shapefile that needs merging}

\item{shape2}{The second shapefile that needs merging}

\item{zone}{The UTM zone of the shapefile/study site}

\item{plot}{Do you want to plot/visualize the merged site?}

\item{save}{Do you want to save the merged site to the working directory?}
}
\value{
A new shapefile resulting from combining two separate shapefiles
}
\description{
Merge two shapefiles together
}
\examples{
#Read in two site shapefiles needing to be merged
shape1 <- readOGR(dsn = paste0("./shape1.shp"))
shape2 <- readOGR(dsn = paste0("./shape2.shp"))

#Select required transect parameters and run function
shape_merge <- site_merge(shape1, shape2, 1, plot = T, save = T)

}
