% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Start_end.R
\name{start_end}
\alias{start_end}
\title{Start end}
\usage{
start_end(lines, zone, plot, save)
}
\arguments{
\item{lines}{A SpatialLines object delineating transect lines}

\item{zone}{The UTM zone of the study site}

\item{plot}{Do you want to plot/visualize the start and end points?}

\item{save}{Do you want to save the SpatialPoints to the working directory?}
}
\value{
A SpatialPoints object showing start and end locations for LTDS lines
}
\description{
Delineates start and end coordinates for LTDS transect lines
}
\examples{
#Read in two site shapefiles needing to be merged
shape1 <- readOGR(dsn = paste0("./shape1.shp"))
shape2 <- readOGR(dsn = paste0("./shape2.shp"))

#Select required transect parameters and run function
shape_merge <- site_merge(shape1, shape2, 1, plot = T, save = T)

}
