% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vegetation_plots.R
\name{veg_plot}
\alias{veg_plot}
\title{Vegetation plots}
\usage{
veg_plot(lines, intv, plot, save)
}
\arguments{
\item{lines}{A SpatialLines object delineating transect lines}

\item{intv}{The interval/distance in meters between plots}

\item{plot}{Do you want to plot/visualize the start and end points?}

\item{save}{Do you want to save the SpatialPoints to the working directory?}
}
\value{
A SpatialPoints object showing start and end locations for LTDS lines
}
\description{
Delineates locations for collecting vegetation obstruction data
}
\examples{
#Read in SpatialLines object showing transect lines
transect.lines <- readOGR(dsn = paste0("./Transect_lines.shp"))

#Input transect lines, and delineate distance between vegetation plots
veg.plot <- veg_plot(transect.lines, 75, plot = T, save = T)

}
