% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Start_end.R
\name{start_end}
\alias{start_end}
\title{Start end}
\usage{
start_end(lines, zone, plot, save)
}
\arguments{
\item{lines}{A SpatialLines object delineating transect lines}

\item{zone}{The UTM zone of the study site}

\item{plot}{Do you want to plot/visualize the start and end points?}

\item{save}{Do you want to save the SpatialPoints to the working directory?}
}
\value{
A SpatialPoints object showing start and end locations for LTDS lines
}
\description{
Delineates start and end coordinates for LTDS transect lines
}
\examples{
#Read in SpatialLines object showing transect lines
transect.lines <- readOGR(dsn = paste0("./Transect_lines.shp"))

#Input transect lines, and delineate UTM zone
transect.points <- start_end(transect.lines, 1, plot = T, save = T)

}
