% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transect_placement.R
\name{tran_place}
\alias{tran_place}
\title{Transect placement}
\usage{
tran_place(tran.dist, pil.perc, direction, site.poly, zone, plot)
}
\arguments{
\item{tran.dist}{The required distance between transects in meters}

\item{pil.perc}{The decimal percentage of overall transects required for pilot surveys}

\item{direction}{The cardinal direction of the transects}

\item{site.poly}{A shape file used to clip the transect lines to fit the site}

\item{zone}{The UTM zone of the shapefile/study site}

\item{plot}{Do you want the transect placement to be plotted against the site?}
}
\value{
A SpatialLines object delineating transect placement
}
\description{
Place transects across site in preparation for LTDS surveys
}
\examples{
#Read in site shapefile file
site.poly <- readOGR(dsn = paste0("./site.poly.shp"))

#Select required transect parameters and run function
transect.lines <- tran_place(50, 0.5, "angle", site.poly, 1, plot = T)

}
